# src/docs/openapi/api.yaml
openapi: 3.0.3
info:
  title: GitHub User Analyzer API
  description: GitHub User Analyzer API
  version: 1.0.0
tags:
  - name: GitHub User Details
    description: GitHub User Details description
paths:
  '/api/users':
    get:
      tags:
        - GitHub User Details
      summary: Returns the list of all github users.
      description: Returns the list of all users with requested pagination parameters.
      operationId: getAllUsers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  '/api/users/{username}':
    get:
      tags:
        - GitHub User Details
      summary: Returns the list of github users with partially matching username.
      description: Returns the list of github users with partially matching username.
      operationId: getUsersByPartialUsername
      parameters:
        - $ref: '#/components/parameters/Username'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  '/api/users/{company}':
    get:
      tags:
        - GitHub User Details
      summary: Returns the list of github users from same company.
      description: Returns the list of users from the mentioned company.
      operationId: getUsersByCompany
      parameters:
        - $ref: '#/components/parameters/Company'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  '/api/users/{location}':
    get:
      tags:
        - GitHub User Details
      summary: Returns the list of github users from same location.
      description: Returns the list of users from the mentioned location.
      operationId: getUsersByLocation
      parameters:
        - $ref: '#/components/parameters/Location'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
components:
  parameters:
    Username:
      in: path
      name: username
      description: user's login
      schema:
        type: string
        default:
      required: true
    Location:
      in: path
      name: location
      description: user's location
      schema:
        type: string
        default:
      required: true
    Company:
      in: path
      name: company
      description: user's company
      schema:
        type: string
        default:
      required: true
    Page:
      in: query
      name: page
      description: page number
      schema:
        type: integer
        default: 1
      required: false
    Size:
      in: query
      name: size
      description: size per page
      schema:
        type: integer
        default: 10
      required: false
  schemas:
    OkResponse:
      type: array
      example: [
        {
          "login": "mary-nazaretyan",
          "id": 45001624,
          "node_id": "MDQ6VXNlcjQ1MDAxNjI0",
          "avatar_url": "https://avatars.githubusercontent.com/u/45001624?v=4",
          "gravatar_id": "",
          "url": "https://api.github.com/users/mary-nazaretyan",
          "html_url": "https://github.com/mary-nazaretyan",
          "followers_url": "https://api.github.com/users/mary-nazaretyan/followers",
          "following_url": "https://api.github.com/users/mary-nazaretyan/following{/other_user}",
          "gists_url": "https://api.github.com/users/mary-nazaretyan/gists{/gist_id}",
          "starred_url": "https://api.github.com/users/mary-nazaretyan/starred{/owner}{/repo}",
          "subscriptions_url": "https://api.github.com/users/mary-nazaretyan/subscriptions",
          "organizations_url": "https://api.github.com/users/mary-nazaretyan/orgs",
          "repos_url": "https://api.github.com/users/mary-nazaretyan/repos",
          "events_url": "https://api.github.com/users/mary-nazaretyan/events{/privacy}",
          "received_events_url": "https://api.github.com/users/mary-nazaretyan/received_events",
          "type": "User",
          "site_admin": false,
          "name": "Mary Nazaretyan",
          "company": null,
          "blog": "",
          "location": null,
          "email": null,
          "hireable": null,
          "bio": null,
          "twitter_username": null,
          "public_repos": 9,
          "public_gists": 0,
          "followers": 0,
          "following": 0,
          "created_at": "2018-11-13T11:18:18Z",
          "updated_at": "2022-01-21T14:45:35Z"
        },
        {
          "login": "aramlog",
          "id": 3672854,
          "node_id": "MDQ6VXNlcjM2NzI4NTQ=",
          "avatar_url": "https://avatars.githubusercontent.com/u/3672854?v=4",
          "gravatar_id": "",
          "url": "https://api.github.com/users/aramlog",
          "html_url": "https://github.com/aramlog",
          "followers_url": "https://api.github.com/users/aramlog/followers",
          "following_url": "https://api.github.com/users/aramlog/following{/other_user}",
          "gists_url": "https://api.github.com/users/aramlog/gists{/gist_id}",
          "starred_url": "https://api.github.com/users/aramlog/starred{/owner}{/repo}",
          "subscriptions_url": "https://api.github.com/users/aramlog/subscriptions",
          "organizations_url": "https://api.github.com/users/aramlog/orgs",
          "repos_url": "https://api.github.com/users/aramlog/repos",
          "events_url": "https://api.github.com/users/aramlog/events{/privacy}",
          "received_events_url": "https://api.github.com/users/aramlog/received_events",
          "type": "User",
          "site_admin": false,
          "name": "Aram Kirakosyan",
          "company": null,
          "blog": "",
          "location": "Yerevan",
          "email": null,
          "hireable": null,
          "bio": null,
          "twitter_username": null,
          "public_repos": 4,
          "public_gists": 0,
          "followers": 1,
          "following": 0,
          "created_at": "2013-02-22T19:47:24Z",
          "updated_at": "2021-07-22T18:09:00Z"
        } ]